+++ src/app-shared/core/core.module.ts
@@ -11,11 +11,11 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ModuleWithProviders, NgModule, Provider, Type } from '@angular/core';
 import { FormControlErrorsComponent } from './components/form-control-errors/form-control-errors.component';
-import { IAccountService } from './interface/account.service';
+import { IAccountService } from './interface/account.interface';
 import { IApiEndpointBuilderService } from './interface/api-endpoint-builder.interface';
-import { INotificationService } from './interface/notification.service';
+import { INotificationService } from './interface/notification.interface';
 import { IStringTemplateService } from './interface/string-template.interface';
-import { ITranslationService } from './interface/translation.service';
+import { ITranslationService } from './interface/translation.interface';
 import { IWebRequestService } from './interface/web-request.interface';
 import { TranslatePipe } from './pipes/translate.pipe';
 import { AccountService } from './service/account.service';
