+++ src/app-shared/core/index.ts
@@ -13,7 +13,7 @@ export { SerializableDataModel } from './data/serializable-data-model';
 export { User } from './data/user.dm';
 export { IAccountCredentials } from './interface/account-credentials';
 export { IAccountInformation } from './interface/account-information';
-export { IAccountService } from './interface/account.service';
+export { IAccountService } from './interface/account.interface';
 export { Dictionary } from './interface/dictionary';
-export { INotificationConfig, INotificationService } from './interface/notification.service';
-export { ITranslationService, TranslationPhrase, TranslationPhraseArgs } from './interface/translation.service';
+export { INotificationConfig, INotificationService } from './interface/notification.interface';
+export { ITranslationService, TranslationPhrase, TranslationPhraseArgs } from './interface/translation.interface';
