+++ src/app-shared/core/interface/translation.interface.ts
