+++ src/app-shared/core/service/notification.service.spec.ts
@@ -6,7 +6,7 @@
 
 import { inject, TestBed } from '@angular/core/testing';
 import { createTranslationServiceStub } from 'app-shared/test-utils';
-import { ITranslationService } from '../interface/translation.service';
+import { ITranslationService } from '../interface/translation.interface';
 import { NotificationService } from './notification.service';
 
 describe('NotificationService', () => {
